Refactor CandidateService adding CandidateRepository
