+++ backend/src/domain/repositories/CandidateRepository.ts
@@ -1,0 +1,6 @@
+import { Candidate } from '../models/Candidate';
+
+export interface CandidateRepository {
+  save(candidate: Candidate): any;
+//   findOne(id: number): Promise<Candidate | null>;
+}
