--- backend/src/presentation/controllers/candidateController.ts
@@ -1,18 +1,0 @@
-import { Request, Response } from 'express';
-import { addCandidate } from '../../application/services/candidateService';
-
-export const addCandidateController = async (req: Request, res: Response) => {
-    try {
-        const candidateData = req.body;
-        const candidate = await addCandidate(candidateData);
-        res.status(201).json({ message: 'Candidate added successfully', data: candidate });
-    } catch (error: unknown) {
-        if (error instanceof Error) {
-            res.status(400).json({ message: 'Error adding candidate', error: error.message });
-        } else {
-            res.status(400).json({ message: 'Error adding candidate', error: 'Unknown error' });
-        }
-    }
-};
-
-export { addCandidate };
