+++ backend/src/routes/candidateRoutes.ts
@@ -1,12 +1,18 @@
 import { Router } from 'express';
-import { addCandidate } from '../presentation/controllers/candidateController';
+import { CandidateService } from '../application/services/candidateService';
+import { PrismaCandidateRepository } from '../infrastructure/repositories/PrismaCandidateRepository';
+import { PrismaClient } from '@prisma/client';
 
 const router = Router();
+const prisma = new PrismaClient();
 
 router.post('/', async (req, res) => {
   try {
     // console.log(req.body); //Just in case you want to inspect the request body
-    const result = await addCandidate(req.body);
+    const candidateData = req.body;
+
+    const candidateService = new CandidateService(new PrismaCandidateRepository(prisma));
+    const result = await candidateService.addCandidate(candidateData);
     res.status(201).send(result);
   } catch (error) {
     if (error instanceof Error) {
